alter groupedEventsCard component
